Teach LockFile how to append to an existing file
In some cases we may want to safely append to a file by copying the
current file content into the temporary file, and then leaving the
temporary file available for further writing by the caller.  This is
technique is really only suitable for smaller files, such as the text
based MERGE_HEAD or FETCH_HEAD.  Binary pack files would not be a good
candidate to append onto with this strategy.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>