Correct IndexPack.renamePack to honor standard pack creation rules
C Git always inserts the .pack file into objects/pack before the
.idx file.  This rule ensures that both files are present in the
pack directory before a runtime process tries to open one of them.
This pack insert rule is based upon the pack read rule, which says
a runtime should only open the ".idx" file, and if the ".idx" is
available then the runtime may be able to read the ".pack" too.

We also never attempt to replace an existing pack file.  If the
final pack exists on disk already under this pack name we can
assume it contains the same set of objects.  We never want to
replace known good local objects with objects received from the
network.  This is a simple safety measure to avoid corruption.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>