Create a pluggable SSH connection factory
SSH support is mostly implemented through a pluggable factory,
allowing a different implementation of the factory to be supplied
by an IDE, such as in the case of Eclipse where the base Eclipse
platform has knowledge of JSch and its configuration for this
end-user.  By deferring most of the user interaction issues to the
factory we are able to make an SSH transport that is common for
all uses (command line, AWT, SWT).

When we are inside of Eclipse we really want to use the user's own
SSH keys and such as configured within the workbench.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>