Trap EOF during fetch connection initialization as "not found"
If a Git client asks a remote peer to run git-upload-pack on a
directory that is not a Git repository we get nothing at all in
return from the git-upload-pack side of the connection; instead
we get EOF while waiting for the length part of the first ref's
packet line.  This is not very user friendly, but is meant as a
way to prevent leaking information from the remote side to an
unfriendly client.

We detect this case and translate it to a TransportException,
using a slightly better error message for the end-user.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>