Delete the prior FetchClient implementation
Now that we have the Transport API laid down to support both push
and fetch, as well as all of the different protocols that C Git
can actually implement, we do not need the previous FetchClient
contributed by Robin.  The implementation in FetchClient was (in
part) stolen to create PackFetchConnection earlier, as that is
the support necessary for a native object transfer, which is all
that FetchClient was capable of doing.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>