Make PackIndex public and expose PackIndex.hasObject
Some application code, such as "dumb" commit walkers that perform
fetching over non-Git aware protocols like HTTP or FTP require an
efficient means to read a pack's .idx file and then test if there
is an occurrence of an object stored within that pack.  The test
frequently gets used by application code to determine if it needs
to download a pack file.

Indexes are generally smaller and thus cheaper to transport over
the dumb protocol, so they are the better method to determine an
object existence test.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>