Allow UnpackedObjectLoader to be called on an already read object
Some applications (e.g. a dumb protocol based fetch process) may
have loaded a loose object over a network stream into memory and
needs to verify the SHA-1 of the content prior to writing out the
already compressed content to disk.  By allowing a loader to be
created for a preloaded compressed object representation we can
allow that application code to reuse our parsing logic and also
to reuse our SHA-1 name computation logic, to verify the object
is what it claims to be.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>