Translate ssh://user@host/~foo/bar to be relative to home directory
If we get a URL of the form "ssh://user@host/~foo" what the user
really is asking for is "bar" inside of foo's home directory, but
they are required to start the path with a "/" to separate it from
the hostname.

We must permit the remote shell to evaluate ~/ or ~who/ in order
to translate to the designated user's home directory, but for the
remainder of the path we must escape it from the shell to protect
whitespace or metacharacters.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>