Add dumb transport support for sftp:// style URLs
We now support dumb transport over the sftp protocol, which is the
file transfer protocol that is built into SSH2.  The protocol is
commonly available on many systems that have web hosting, but not
Git hosting.  These systems sometimes require publishing content
only through SSH, and recommend sftp clients to their customers
for account management.

Fetching from such a repository is of relatively little value, as
fetching over the unencrypted (and more lightweight) HTTP transport
will be faster.  However we do want to offer push support over this
protocol in the future, so implementing fetch support for it via
the generic dumb walker would at least make our support for this
protocol symmetrical.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>