Honor refspecs passed to our command line Fetch client
While testing and experimenting with the http based fetch support
I missed having refspecs on the command line, especially for the
single-shot fetch/pull case that is so commonly used in C Git.
Its easy enough to support here; we just need to create the spec
objects in a list out of the remaining arguments.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>