Paper bag fix RevWalk.parseAny to always parse objects
If the caller has asked us to parse an object from an ObjectId the
object they are requesting may already be in our cache, as it could
be a tree or a parent commit for a commit we have already parsed.
Since the object is in our cache we don't bother to open it from the
object database to examine its type, but it may not be parsed yet.
We need to parse it if it isn't.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>