Change UnpackedObjectLoader(byte[]) to only throw CorruptObjectException
This constructor can only throw a CorruptObjectException as it never
needs to read from a disk based representation of the object.  This
can simplify callers who use this special constructor as they only
need to handle the corrupt object case and not a generic IO failure.

Further we cannot rely on getId() in this constructor as the id is
null and the data failed to decompress.  The attempt to compute the
id from the decompressed bytes would result in a NullPointerException
as the computation tries to dereference the data.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>