Define static ASCII encoded type strings for ObjectId computation
We frequently need to compute the ObjectId of something, and to do
that we must embed the type name as an ASCII string into the header
as part of the digest calculation.  This means we are often building
these ASCII encoded strings, but there's only 4 possible values in
Git so it doesn't make sense to keep redoing the String->byte[] on
every ObjectId computation.  Instead we store them as static byte[]
and pray nobody will be foolish enough to edit our constant data.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>