Allow unthrown exceptions to be declared as thrown in jgit methods
The following is not a compile error:

  class A {
    void parse(RevObject o) throws IOException {
      o.flags |= PARSED;
    }
  }

  class B extends A {
	Repository local;

	@Override
    void parse(RevObject o) throws IOException {
	  super.parse(o);
	  final ObjectLoader ldr = local.openObject(o);
	}
  }

If we write a method with a default implementation but an overidden
implementation of the same method needs to throw a checked exception,
that checked exception must get declared in the initial declaration
of the method.  This should not be an error.

It may be useful as a warning, to signal that perhaps the throws
declaration is unnecessary, so I am not ignoring it outright.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>