Implement FileMode.fromBits utility in TreeWalk
I'm finding that we need to convert the raw mode bits supplied by
TreeWalk into a FileMode in a number of places, usually so that we
can then obtain the OBJ_* type constant that the FileMode refers
to in a tree entry.  Creating a utility for this translation in
TreeWalk just saves us some redundant code in applications.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>