Defer creation of RevWalk in RevWalk based command tools
We need to first get through the bulk of our options parsing so
we know what sort of RevWalk we might need to create to handle
the request submitted by the caller on the command line.  Right
now we don't have another type to select from, but that will not
be the case once we start to handle --objects.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>