Always initialize TreeWalk with a single empty tree
By loading a single empty tree iterator into the tree walk's
initial list of trees we can avoid array index out of bound
exceptions if the application tries to call next() prior to
actually giving us a tree to iterate.  This can simplify a
few application cases where we may need to advance an open
tree walk, or load a new tree into one if we are already at
the end and have no more entries.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>