Only request advertised include-tags if we want tags
If we are using the native git pack transport and the other side
has advertised that it can include tags for us if the peeled ref
points to an object it has already packed to send us we may not
want it to perform that packing for us if we are not going to be
fetching annotated tags at this time.  So we now only enable the
include-tags protocol option if some form of tag fetching was
set on the transport instance we are being created from.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>