Allow FetchConnection users to see if tag following was used
If the remote side was able to pack tags for us and send them
(such as a recent git-upload-pack can do) applications may be
able to avoid a second native connection to setup the missing
tag objects that they do not yet have.  However looking to see
if that is the case is only worthwhile if the connection was
able to use this feature.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>