Skip pack-*.idx which has no corresponding pack-*.pack file
The HTTP transport in C Git will download a pack-*.idx file from the
remote to see what is contained in the pack, but it does not get the
matching pack-*.pack if the object(s) are already available locally.
As these are written directly to the .git/objects/pack directory we
must skip over the useless index files when setting up our list of
available PackFile instances.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>