Add TODO marker about SWTWalk not disposing in HistoryPage
We shouldn't need to dispose of the SWTWalk here.  The walker has
a bug and is not properly resetting itself when HEAD is reset to
HEAD^1 or some other value.  Robin has worked around the issue by
forcing a reset for any change in HEAD as part of d5a698391f17,
but that is only a workaround as it removes any caching that the
SWTWalk might be able to perform.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>