Paper bag fix quoting for SSH transport commands
Not all Git-over-SSH servers run a Bourne shell on the remote side
to evaluate the command we are sending.  Some servers run git-shell,
which will fail to execute git-upload-pack if we feed it a quoted
string for the name git-upload-pack.

Testing concludes that git-shell requires the command name to never
be quoted, and the argument name to always be single quoted.  As
this is a long-standing behavior in the wild jgit needs to conform,
as git-shell and all git-shell work-a-likes such as gitosis may be
following the same convention.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>