Support push over the sftp:// dumb transport
jgit now allows users to push changes over the sftp:// protocol,
taking advantage of the SFTP client available as part of JSch to
(more or less) safely update the remote repository.

Since locking is not available over SFTP this is not suitable for
use with concurrent pushes, but is safe with concurrent fetches
during a single push.  This is sufficient support to safely update
a Git repository published over HTTP where the only means of making
changes is through SSH/SFTP and the remote side does not make Git
available through the shell, or does not offer direct shell access.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>