Support 'git upload-pack' and 'git receive-pack' over SSH
Within the next 6 months C git clients will begin asking remote
servers for 'git $command' rather than 'git-$command' when using the
SSH transport.  This change is to allow the C git programs to be
removed from the user's $PATH, leaving only the git wrapper binary.

For the first 6 months after C git 1.6.0 gets released clients will
continue to ask for 'git-$command' but users may change that behavior
by specifically asking for 'git $command' in remote.$name.uploadpack
or remote.$name.receivepack.  Later clients (including jgit) will
change to ask for 'git $command' by default.

If we are asking for 'git $command' we cannot quote this as a single
command with a space in the path.  We split on the whitespace and
quote both sides (if necessary) to protect the strings from the shell.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>