Use 'git upload-pack' for local transport and not 'git-upload-pack'
Since Git 1.6.0 and later is heading towards moving the dash form
of commands out of the user's $PATH and into the libexec directory
we may not have 'git-upload-pack' or 'git-receive-pack' in $PATH
when the JRE tries to startup C Git for a local transport.  Instead
we should use the git wrapper, as it can select the right libexec
directory and start the correct helper program, or perform the task
internally if it is a builtin command.

In the future we should eliminate this reliance on C Git for any
sort of local transport operation.  We have direct filesystem IO
available to read from the source repository, so there is little
execuse for us to be invoking C Git for this special case.  Until
we fix it, we should at least try to ensure it works for any user.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>